__main__/update: Report errors instead of raising exceptions
'dfu-util' is an external util that may or may not be installed on the
host running the tool.

This change captures exceptions that happen when the tool is not
available on the system or if the tools fails with ann return code and
presents the user with a human-readable error message instead of a
python-stacktrace.