Handle that dev.port_numbers can be None
With libusb on Windows the dev.port_numbers attribute is not set for
USB-Devices.
With this change this case is handled: Paths will always be empty on
Windows.
This means that selecting an USB-Mux by path will not work on this
platform. (But at least we will not raise an exception.)

Signed-off-by: Chris Fiege <cfi@pengutronix.de>